Create Change From Baseline Over Time Table

create change from baseline over time table, table can be ordered by param, treatment group and visit

Usage

qc_chgfb(
  input,
  val = "AVAL",
  chg = "CHG",
  rowvar = c("PARAM", "TRT01P", "AVISIT"),
  stats_list = c("N", "Mean", "SD", "Median", "Min", "Max"),
  max_digit = 2,
  keep = TRUE
)

Arguments

input

input dataframe

val

measurement variable

chg

change from baseline measurement variable

rowvar

row variable (can be a set of two variables such as c("TRT01P", "AVISIT") or three variables such as c("PARAM", "TRT01P", "AVISIT"))

stats_list

stats variables to display. Accepted values are c("N", "Mean", "SD", "Median", "Min", "Max", "CV", "Base_mean") and the order of variables matters

max_digit

largest number of digit to display

keep

if = TRUE, keep all levels. Default = TRUE

Value

change from baseline table

Details

The number of decimal places the results will keep depends on a variable "digit" in the input dataset. If digit variable does not exist in input dataset, the number of decimal places will depend on the number of decimal places of AVAL/CHG

Examples

adsl <- data.frame(
  USUBJID = 1:10,
  TRT01P = factor(sample(c("A", "B", "C"), 10, replace = TRUE)))
  
param <- data.frame(PARAM = c("Test1", "Test2"))
visit <- data.frame(AVISIT = c("Baseline", "Visit1", "Visit2"))

adlb0 <- merge(adsl, param)
adlb0$BASE <- sample(1:100, 20)

adlb <- merge(adlb0, visit)
adlb$AVAL <- sample(1:100, 60)
adlb$CHG <- ifelse(adlb$AVISIT == "Baseline", NA, adlb$AVAL - adlb$BASE)
adlb$digit <- ifelse(adlb$PARAM == "Test1", 0, 2)

tab1 <- qc_chgfb(adlb, "AVAL", "CHG", rowvar = c("PARAM", "TRT01P", "AVISIT"), 
                 max_digit = 0, keep = FALSE)
tab1
#> # A tibble: 26 × 13
#>    row_text   N.x Mean.x SD.x  Median.x Min.x Max.x   N.y Mean.y SD.y  Median.y
#>    <chr>    <int> <chr>  <chr> <chr>    <chr> <chr> <int> <chr>  <chr> <chr>   
#>  1 Test1       NA NA     NA    NA       NA    NA       NA NA     NA    NA      
#>  2 A           NA NA     NA    NA       NA    NA       NA NA     NA    NA      
#>  3 Baseline     4 79.0   22.70 85.0     47    99       NA NA     NA    NA      
#>  4 Visit1       4 51.8   45.76 52.5     4     98        4 7.5    53.57 6.5     
#>  5 Visit2       4 43.3   19.82 40.5     24    68        4 -1.0   50.26 -5.5    
#>  6 B           NA NA     NA    NA       NA    NA       NA NA     NA    NA      
#>  7 Baseline     2 57.5   27.58 57.5     38    77       NA NA     NA    NA      
#>  8 Visit1       2 58.5   43.13 58.5     28    89        2 24.0   4.24  24.0    
#>  9 Visit2       2 37.0   41.01 37.0     8     66        2 2.5    2.12  2.5     
#> 10 C           NA NA     NA    NA       NA    NA       NA NA     NA    NA      
#> # ℹ 16 more rows
#> # ℹ 2 more variables: Min.y <chr>, Max.y <chr>