Create Demographic Rows

Function to create demographic rows

Usage

qc_demo(
  input,
  colvar = "TRT01P",
  stats_list = c("Mean_SD", "Median", "Range"),
  var_list,
  drop_var_list = NULL,
  max_digit = 2
)

Arguments

input

input dataframe

colvar

column variable

stats_list

stats variables to display. Accepted values are c("Mean_SD", "Mean", "SD", "Median", "Range", "Min", "Max", "Geo_mean", "Geo_CV", "Geo_CL") and the order of variables matters

var_list

variable list in demo table. Please see details.

drop_var_list

Variable that don't keep all levels in the output

max_digit

maximum number of decimal place to report

Value

dataframe with demographic rows

Details

var_list should be a dataframe with desired variable names as column names, the label of each variable display in tables as the first row, and optionally, the number of decimal places to keep in the output as the second row. If the label row of a variable is empty, the row text and N row for this variable will be missing.

Examples

age <- sample(18:65, 10, replace = TRUE)

adsl <- data.frame(
  USUBJID = 1:10,
  TRT01P = factor(sample(c("A", "B", "C"), 10, replace = TRUE)),
  AGE = age,
  AGEGR1 = ifelse(age < 45, "< 45", ">= 45"),
  SEX = factor(sample(c("Female", "Male"), 10, replace = TRUE)))
  
### Create variable list based on DPS and assign labels to them.
### Leave "" For variables that concatenate to the corresponding categorical variables
var_list <- data.frame(
  AGE = "Age, years",
  AGEGR1 = "",
  SEX = "Sex"
  )

tab1 <- qc_demo(adsl, colvar = "TRT01P", var_list = var_list)
tab1
#>      row_text            A           B            C
#> 1  Age, years         <NA>        <NA>         <NA>
#> 2           N            5           2            3
#> 3   Mean (SD) 41.6 (18.88) 57.5 (0.71) 44.0 (15.39)
#> 4      Median         44.0        57.5         48.0
#> 5       Range     (18; 65)    (57; 58)     (27; 57)
#> 6        < 45    3 (60.0%)        <NA>    1 (33.3%)
#> 7       >= 45    2 (40.0%)  2 (100.0%)    2 (66.7%)
#> 8         Sex         <NA>        <NA>         <NA>
#> 9           N            5           2            3
#> 10     Female    3 (60.0%)   1 (50.0%)    2 (66.7%)
#> 11       Male    2 (40.0%)   1 (50.0%)    1 (33.3%)